Title: 괄호 쌍 만들어주기, Time: 201ms, Memory: 14MB, Status: Passed - Codetree

+++ 250528/괄호 쌍 만들어주기/pair-parentheses.java
@@ -18,12 +18,12 @@ public class Main {
             }
         }
 
-        int total = 0;
+        long total = 0;
 
         for (int i = 1; i<A.length(); i++) {
             
             if (k[i] == '(' && k[i-1] == '(') {
-                total += sum[i];
+                total += (long)sum[i];
             }
         }
